Use Clickhouse readonly user. Improve function names and remove unused code

+++ lib/sanbase_web/graphql/schema/transaction_types.ex
@@ -10,7 +10,7 @@ defmodule SanbaseWeb.Graphql.TransactionTypes do
   object :transaction do
     field(:datetime, non_null(:datetime))
     field(:trx_hash, non_null(:string))
-    field(:trx_value, non_null(:decimal))
+    field(:trx_value, non_null(:float))
     field(:from_address, non_null(:string))
     field(:to_address, non_null(:string))
   end
